Give cpp-use-cases.md a more recent gtest example
Most importantly, the old example doesn't work; the new one does. Since this is the most canonical document I could find for how to use gtest with bazel, that seemed like the important bar.

The old example refers to a version of gtest that's a few revisions behind, so I imagine the git_repository rule was added since this doc was written. It's true that this shorter version teaches the reader much less detail about the mechanics of external dependencies. If that's the goal, feel free to reject this proposal. But I suspect that these lessons are also out of date, and moreover those details are the proper subject of a different page, like https://docs.bazel.build/versions/4.0.0/external.html.